requirements.txt
python-dotenv
gigachain-community
gigachain-cli
